app/requirements.txt
asgiref
certifi
charset-normalizer
cryptography
defusedxml
mysqlclient
Django
idna
oauthlib
Pillow
pycparser
PyJWT
python-dotenv
python3-openid
pytz
requests
requests-oauthlib
social-auth-app-django
social-auth-core
sqlparse
urllib3
